refactor(cost): house caps compliance (1500-LOC files, 100-line fns)
The ratchets bit and forced the better structure (the known pattern):
- runtime: budget_tests -> tests/budget_gate.rs (integration) ·
  dispatch_wave extracted from run() · ledger.debit_ok(&DispatchOk)
  absorbs the leaf-debit plumbing · race_budget extracted from
  attempt_loop
- cli: run/budget.rs submodule owns the --max-cost-usd preflight
  (floor_refusal pure + tested) · trace_sink extracted
- catalog: cache-aware pricing-math tests -> data/pricing_math_tests.rs
- display: apply_task_completed extracted from the fold's apply
- verb-agent: finished() owns the Finished-event + spend decoration

Zero behavior change — all 8 CI ratchets green.
