+++ crates/nika-cli/src/verbs/run/mod.rs
@@ -36,6 +36,7 @@ pub use sink::{FoldSink, JsonSink, RenderMode};
 use sink::{TraceNote, surface_trace};
 pub use stamp::SystemStamper;
 
+mod budget;
 mod scope;
 use scope::scope_to_task;
 pub(crate) use source_id::{lf_normal_form, sha256_hex};
@@ -116,16 +117,14 @@ pub fn run(
     let (source, wf, report) = match crate::verbs::load_checked_with_source(file) {
         Ok(pair) => pair,
         Err(out) => {
-            // Pre-run diagnostics obey the export contract too: in machine
-            // mode they go to stderr so a `capture: stdout` consumer never
-            // mistakes the "cannot read" text for the JSON result.
+            // Machine mode: diagnostics ride stderr so a `capture: stdout`
+            // consumer never mistakes them for the JSON result.
             emit_diagnostic(&refusal_text(&out), output_json);
             return out.code;
         }
     };
 
-    // ── `--task` scope + clean gate + `--var` overrides (all before
-    //    any effect — the whole operator-input preflight) ──
+    // ── `--task` scope + clean gate + `--var` overrides (pre-effect) ──
     let (wf, report) =
         match scoped_clean_gate(wf, report, task_filter, file, json, theme, output_json) {
             Ok(pair) => pair,
@@ -141,24 +140,9 @@ pub fn run(
         return render_dry_run(file, theme.ascii);
     }
 
-    // ── `--max-cost-usd` preflight — BEFORE any spend ────────────────
-    // Refuse when the STATIC floor (cheapest path · gates closed ·
-    // first-try · the unavoidable exposure `nika check` computes)
-    // already exceeds the budget; warn loud when the ceiling cannot
-    // bound everything (the budget gates METERED spend only).
-    if let Some(budget) = max_cost_usd {
-        if let Some(refusal) = budget_floor_refusal(report.cost.min_path_total_usd, budget) {
-            emit_diagnostic(&refusal, output_json);
-            return exit::FILE;
-        }
-        if report.cost.has_unbounded {
-            let unbounded = report.cost.tasks.iter().filter(|t| t.usd.is_none()).count();
-            eprintln!(
-                "⚠ --max-cost-usd {budget}: {unbounded} task(s) have no static ceiling \
-                 (no token bound · unknown iterations · unpriced model) — the budget \
-                 bounds METERED spend only; local/mock work never trips it"
-            );
-        }
+    // ── `--max-cost-usd` preflight — BEFORE any spend (budget.rs) ──
+    if let Err(code) = budget::preflight(&report, max_cost_usd, output_json) {
+        return code;
     }
 
     // ── `--resume` / `--answer` (ADR-099) — plan + answers up front ──
@@ -167,8 +151,7 @@ pub fn run(
         Err(code) => return code,
     };
     // ADR-099: the pause rider binds to the NON-INTERACTIVE machine
-    // surfaces only (`--json` · `--output json`); human TTY/plain
-    // surfaces keep today's PROMPT-001 contract untouched.
+    // surfaces only — human TTY/plain keep the PROMPT-001 contract.
     let pause_on_prompt = json || output_json;
 
     // ── Compose the production runtime (real seams · env keys) ──────
@@ -192,13 +175,7 @@ pub fn run(
         Ok(rt) => rt,
         Err(code) => return code,
     };
-    // The run journal (spec §3.3) — composed HERE like the other seams;
-    // `execute` receives the sink, never a flag.
-    let trace = if no_trace_file {
-        TraceFileSink::disabled()
-    } else {
-        TraceFileSink::new(TRACE_DIR)
-    };
+    let trace = trace_sink(no_trace_file);
     rt.block_on(execute(
         &runtime,
         (file, &wf),
@@ -990,45 +967,13 @@ fn outputs_json_line(outputs: &BTreeMap<String, Value>) -> String {
 /// contract · `capture: stdout` composition), stdout in the human modes.
 /// In machine mode the failure ALSO lands on stdout as the `{"error":{…}}`
 /// envelope (F6) — the machine surface is self-sufficient, success or not.
-/// The `--max-cost-usd` pre-run gate — `Some(refusal)` when the STATIC
-/// floor (cheapest path · gates closed · first-try: the unavoidable
-/// exposure `nika check` computes) already exceeds the budget. Pure so
-/// the operator-facing gate is unit-testable; a floor AT the budget
-/// passes (spending exactly the budget is not over it).
-fn budget_floor_refusal(floor: f64, budget: f64) -> Option<String> {
-    (floor > budget).then(|| {
-        format!(
-            "refusing to start: the workflow's unavoidable cost floor \
-             ${floor:.6} exceeds --max-cost-usd ${budget:.6} (cheapest \
-             static path · gates closed · first-try) — raise the budget \
-             or trim the workflow (`nika check` shows the envelope)\n"
-        )
-    })
-}
-
-/// The operator-facing budget preflight — pure-fn pinned (F4.2: the
-/// gate the operator actually touches must not ride untested).
-#[cfg(test)]
-#[allow(clippy::expect_used, clippy::unwrap_used, clippy::panic)]
-mod budget_preflight_tests {
-    use super::budget_floor_refusal;
-
-    #[test]
-    fn floor_above_budget_refuses_with_both_numbers() {
-        let msg = budget_floor_refusal(0.000_019, 0.000_001).expect("refuses");
-        assert!(msg.contains("$0.000019"), "floor rides: {msg}");
-        assert!(msg.contains("$0.000001"), "budget rides: {msg}");
-        assert!(msg.contains("refusing to start"), "{msg}");
-        assert!(msg.contains("nika check"), "points at the envelope: {msg}");
-    }
-
-    #[test]
-    fn floor_at_or_under_budget_passes() {
-        // Spending exactly the budget is not over it (mirrors the
-        // ledger's crossing semantics).
-        assert!(budget_floor_refusal(0.05, 0.05).is_none());
-        assert!(budget_floor_refusal(0.0, 0.05).is_none());
-        assert!(budget_floor_refusal(0.0, 0.0).is_none());
+/// The run journal (spec §3.3) — composed like the other seams;
+/// `execute` receives the sink, never a flag.
+fn trace_sink(no_trace_file: bool) -> TraceFileSink {
+    if no_trace_file {
+        TraceFileSink::disabled()
+    } else {
+        TraceFileSink::new(TRACE_DIR)
     }
 }
